End of first possibility TextListener : Wrong way
