--- xGenuini/core/assets/img/auteurs.txt
@@ -1,2 +1,0 @@
-personnages : Ronan Defours
-fond : http://opengameart.org/content/landscape
