+++ xGenuini/core/assets/img/book/auteur.txt
@@ -1,0 +1,2 @@
+AnthonyMyers from http://opengameart.org/
+https://creativecommons.org/licenses/by/3.0/
