+++ xGenuini/core/src/genuini/screens/package-info.java
@@ -1,0 +1,11 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+/**
+ * Package for screen.
+ * Contain the AbstractScreen and sons.
+ */
+package genuini.screens;
